requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
python-multipart==0.0.20
psycopg[binary]==3.2.6
redis==5.2.1
pytest==8.3.5
httpx==0.28.1
email-validator==2.2.0

